% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOMicFunctions.r
\name{Mode}
\alias{Mode}
\alias{cdiff}
\title{Find the Mode (most likely value)}
\usage{
Mode(x, na.rm = FALSE)

cdiff(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
The mode.
\item{cdiff}{The mode.}
\item{x}{A vector.}

A vector.
\item{cdiff}{The inverse of cumsum.}
\item{x}{A vector.}
}
\description{
Finds mode.

Finds the inverse of cumsum.
}
\author{
Dominic Woolf

Dominic Woolf
}
