% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOMicFunctions.r
\name{fT.PrimC}
\alias{fT.PrimC}
\title{PrimC temperature response function}
\usage{
fT.PrimC(temp, method = "Century2", t.ref = 30)
}
\arguments{
\item{Temp}{Temperature in degrees C.}
}
\value{
temperature rate modifier.
}
\description{
Calculates temperature rate modifier
}
\author{
Dominic Woolf
}
