% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOMicFunctions.r
\name{fT.RothC}
\alias{fT.RothC}
\title{RothC temperature response function}
\usage{
fT.RothC(Temp)
}
\arguments{
\item{Temp}{Temperature in degrees C.}
}
\value{
temperature rate modifier.
}
\description{
Calculates temperature rate modifier
}
\author{
Dominic Woolf
}
