% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOMicFunctions.r
\name{g.map}
\alias{g.map}
\alias{get_data}
\title{Initialise Draw a pretty map}
\usage{
g.map(r)

get_data(data_path)
}
\arguments{
\item{data_path}{Path to data.}

\item{pattern}{filename pattern}

\item{full.names}{full names}

\item{recursive}{look in subdirectories too?}
}
\value{
A data.frame for the somic model.

A data.frame for the somic model.
}
\description{
Draws an SOC map

Initialises model inputs from files
}
\note{
This is a very simple function.

This is a very simple function.
}
\author{
Dominic Woolf

Dominic Woolf
}
