% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{somic}
\alias{somic}
\title{Runs the SOMic model}
\usage{
somic(
  soc_data,
  mic_vmax,
  mic_km,
  kdissolution,
  kdepoly,
  kdeath_and_exudates,
  kdesorb,
  ksorb,
  kmicrobial_uptake,
  cue_0,
  mcue,
  mclay,
  clay,
  thickness = 23,
  sat = 1,
  max_tsmd = 0,
  use_atsmd = 0L,
  init_spm_14c_age = 0,
  init_ipm_14c_age = 0,
  init_doc_14c_age = 0,
  init_mb_14c_age = 0,
  init_mac_14c_age = 0,
  init_soc_14c_age = NA_real_
)
}
\arguments{
\item{soc_data}{data.frame. See Details for required input columns.}

\item{parameters:}{mic_vmax, mic_km, kdissolution, kdepoly, kdeath_and_exudates, kdesorb, ksorb, kmicrobial_uptake, cue_0, mcue, mclay,}

\item{more_parameters:}{thickness = 23.0, sat = 1.0, max_tsmd = 0.0, use_atsmd = 0, init_spm_14c_age = 0.0, init_ipm_14c_age = 0.0, init_doc_14c_age = 0.0, init_mb_14c_age = 0.0, init_mac_14c_age = 0.0, init_soc_14c_age}
}
\value{
data frame of SOC dynamics
}
\description{
Runs the SOMic model
}
