% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOMicFunctions.r
\name{unrotate}
\alias{unrotate}
\title{Convert rasters to 0 to 360 longitudes}
\usage{
unrotate(x)
}
\arguments{
\item{Sand}{Sand in percent.}

\item{Silt}{Silt in percent.}

\item{Clay}{Clay in percent.}
}
\value{
Unrotated raster.
}
\description{
Provides the inverse function of raster::rotate
}
\author{
Dominic Woolf
}
