% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOMicFunctions.r
\name{wfield}
\alias{wfield}
\title{Estimate field water holding capacity}
\usage{
wfield(sand, silt, clay)
}
\arguments{
\item{Sand}{Sand in percent.}

\item{Silt}{Silt in percent.}

\item{Clay}{Clay in percent.}
}
\value{
Field water holding capacity (m/m).
}
\description{
Calculates field capacity
}
\author{
Dominic Woolf
}
