% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOMicFunctions.r
\name{wmean.bricks}
\alias{wmean.bricks}
\title{Weighted mean of raster bricks}
\usage{
wmean.bricks(b.list, w)
}
\arguments{
\item{b.list}{A list of bricks.}

\item{w}{A vector of weights- same length as \code{b.list}.}
}
\value{
A ratser brick containing the weighted means.
\item{wmean.bricks}{The sum of the squared values.}
\item{b.list}{A list of raster bricks.}
\item{w}{weights (same length as b.list).}
}
\description{
Finds the weighted mean of bricks.
}
\note{
This is a very simple function.
}
\author{
Dominic Woolf
}
