% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOMicFunctions.r
\name{wsat}
\alias{wsat}
\title{Estimate saturated water holding capacity}
\usage{
wsat(sand, silt, clay)
}
\arguments{
\item{Sand}{Sand in percent.}

\item{Silt}{Silt in percent.}

\item{Clay}{Clay in percent.}
}
\value{
saturated water holding capacity (m/m).
}
\description{
Calculates saturation point
}
\author{
Dominic Woolf
}
