% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOMicFunctions.r
\name{wwilt}
\alias{wwilt}
\title{Estimate wilting point}
\usage{
wwilt(sand, silt, clay)
}
\arguments{
\item{Sand}{Sand in percent.}

\item{Silt}{Silt in percent.}

\item{Clay}{Clay in percent.}
}
\value{
Wilting point (m/m).
}
\description{
Calculates wilting point
}
\author{
Dominic Woolf
}
