% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOMicFunctions.r
\name{initialise.somic.data}
\alias{initialise.somic.data}
\title{Initialise Somic data}
\usage{
initialise.somic.data(soc.data)
}
\arguments{
\item{soc.data}{A data frame.}
}
\value{
The data.frame but with more stuff filled in.
\item{soc.data}{The data.frame but with more stuff filled in.}
}
\description{
Initialises model inputs
}
\note{
This is a very simple function.
}
\author{
Dominic Woolf
}
